+++ lib/repositories/grocery_repository.dart
@@ -1,5 +1,5 @@
 import 'package:recipes/data_mapper.dart';
-import 'package:recipes/database.dart';
+import 'package:recipes/database/database.dart';
 import 'package:recipes/models/grocery.dart';
 
 class GroceryRepository {
@@ -8,25 +8,28 @@ class GroceryRepository {
   GroceryRepository({required this.database});
 
   Future<List<Grocery>> getGroceries() async {
-    List<GroceryData> groceriesData = await database.getGroceries();
+    List<GroceryData> groceriesData =
+        await database.groceriesDao.getGroceries();
 
     return DataMapper.groceriesFromData(groceriesData);
   }
 
-  Future<int> addGrocery(Grocery grocery) async => database.addGrocery(grocery);
+  Future<int> addGrocery(Grocery grocery) async =>
+      database.groceriesDao.addGrocery(grocery);
 
   Future<void> updateGrocery(Grocery grocery) async =>
-      database.updateGrocery(grocery);
+      database.groceriesDao.updateGrocery(grocery);
 
   Future<void> deleteGrocery(int id) async {
-    return database.deleteGrocery(id);
+    return database.groceriesDao.deleteGrocery(id);
   }
 
   Future<void> insertOrUpdateGroceries(List<Grocery> groceries) async =>
-      database.insertOrUpdateGroceries(groceries);
+      database.groceriesDao.insertOrUpdateGroceries(groceries);
 
-  Future<void> deleteGroceries() async => database.deleteGroceries();
+  Future<void> deleteGroceries() async =>
+      database.groceriesDao.deleteGroceries();
 
   Future<void> toggleGroceryBought(Grocery grocery) async =>
-      database.toggleGroceryBought(grocery);
+      database.groceriesDao.toggleGroceryBought(grocery);
 }
