+++ lib/repositories/ingredient_repository.dart
@@ -1,4 +1,4 @@
-import 'package:recipes/database.dart';
+import 'package:recipes/database/database.dart';
 import 'package:recipes/models/ingredient.dart';
 
 class IngredientRepository {
@@ -7,17 +7,18 @@ class IngredientRepository {
   IngredientRepository({required this.database});
 
   Future<void> addIngredients(int recipeId, List<Ingredient> ingredientList) {
-    return database.addIngredients(recipeId, ingredientList);
+    return database.ingredientsDao.addIngredients(recipeId, ingredientList);
   }
 
   Future<void> updateRecipeIngredients(
       int recipeId, List<Ingredient> ingredientList) {
-    return database.updateRecipeIngredients(recipeId, ingredientList);
+    return database.ingredientsDao
+        .updateRecipeIngredients(recipeId, ingredientList);
   }
 
   // getRecipeIngredients needed??
 
   Future<void> deleteRecipeIngredients(int recipeId) {
-    return database.deleteRecipeIngredients(recipeId);
+    return database.ingredientsDao.deleteRecipeIngredients(recipeId);
   }
 }
