+++ lib/repositories/recipe_repository.dart
@@ -1,6 +1,6 @@
 import 'package:path_provider/path_provider.dart';
 import 'package:recipes/data_mapper.dart';
-import 'package:recipes/database.dart';
+import 'package:recipes/database/database.dart';
 import 'package:recipes/models/meal_recipe.dart';
 import 'package:recipes/models/recipe_detail.dart';
 import 'package:recipes/models/recipe_list_item.dart';
@@ -12,10 +12,11 @@ class RecipeRepository {
   RecipeRepository({required this.database});
 
   Future<RecipeDetail> getRecipe(int id) async {
-    final recipeData = await database.getRecipe(id);
-    final ingredientsData = await database.getRecipeIngredients(id);
-    final tagsData = await database.getRecipeTags(id);
-    final imageData = await database.getRecipeImage(id);
+    final recipeData = await database.recipesDao.getRecipe(id);
+    final ingredientsData =
+        await database.ingredientsDao.getRecipeIngredients(id);
+    final tagsData = await database.tagsDao.getRecipeTags(id);
+    final imageData = await database.recipesDao.getRecipeImage(id);
 
     final directory = await getApplicationDocumentsDirectory();
 
@@ -24,25 +25,25 @@ class RecipeRepository {
   }
 
   Future<List<RecipeListItem>> getRecipes() async {
-    List<RecipeData> recipeData = await database.getRecipes();
+    List<RecipeData> recipeData = await database.recipesDao.getRecipes();
 
     return DataMapper.recipesFromData(recipeData);
   }
 
   Future<int> addRecipe(RecipeDetail recipe) {
-    return database.addRecipe(recipe);
+    return database.recipesDao.addRecipe(recipe);
   }
 
   Future<void> updateRecipe(RecipeDetail recipe) {
-    return database.updateRecipe(recipe);
+    return database.recipesDao.updateRecipe(recipe);
   }
 
   Future<void> deleteRecipe(int id) {
-    return database.deleteRecipe(id);
+    return database.recipesDao.deleteRecipe(id);
   }
 
   Future<void> toggleFavoriteRecipe(RecipeDetail recipe) {
-    return database.toggleFavoriteRecipe(recipe);
+    return database.recipesDao.toggleFavoriteRecipe(recipe);
   }
 
   // TODO: do searchRecipes parameters need to be required?
@@ -51,16 +52,16 @@ class RecipeRepository {
       required String query,
       required List<Tag> tags,
       required bool favorites}) {
-    return database.searchRecipes(
+    return database.recipesDao.searchRecipes(
         offset: offset, query: query, tags: tags, favorites: favorites);
   }
 
   Future<int> getRecipesCount() {
-    return database.getRecipesCount();
+    return database.recipesDao.getRecipesCount();
   }
 
   Future<List<MealRecipe>> getRecipesById(Set<int> recipeIds) async {
-    List<RecipeData> data = await database.getRecipesById(recipeIds);
+    List<RecipeData> data = await database.recipesDao.getRecipesById(recipeIds);
 
     return DataMapper.mealRecipesFromData(data);
   }
