+++ lib/screens/edit_recipe.dart
@@ -3,14 +3,15 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:get_it/get_it.dart';
 import 'package:recipes/blocs/tags/bloc.dart';
 import 'package:recipes/blocs/tags/events.dart';
-import 'package:recipes/database.dart';
+import 'package:recipes/database/database.dart';
 import 'package:recipes/models/recipe_detail.dart';
 import 'package:recipes/screens/recipe.dart';
 import 'package:recipes/widgets/recipe_form.dart';
 
 Future<int> submitRecipe(BuildContext context, RecipeDetail recipe) async {
-  await GetIt.I<AppDatabase>().updateRecipe(recipe);
+  await GetIt.I<AppDatabase>().recipesDao.updateRecipe(recipe);
   await GetIt.I<AppDatabase>()
+      .ingredientsDao
       .updateRecipeIngredients(recipe.id!, recipe.ingredients);
 
   if (context.mounted) {
