+++ lib/screens/meal_plan.dart
@@ -4,7 +4,7 @@ import 'package:get_it/get_it.dart';
 import 'package:recipes/blocs/meal_plan/bloc.dart';
 import 'package:recipes/blocs/meal_plan/events.dart';
 import 'package:recipes/blocs/meal_plan/state.dart';
-import 'package:recipes/database.dart';
+import 'package:recipes/database/database.dart';
 import 'package:recipes/helpers/number_formatters.dart';
 import 'package:recipes/helpers/add_ingredients_to_groceries.dart';
 import 'package:recipes/models/meal_plan.dart';
@@ -102,7 +102,9 @@ Future confirmMealPlanDelete(BuildContext context, int mealPlanId) async {
             child: const Text('Yes'),
             onPressed: () async {
               try {
-                await GetIt.I<AppDatabase>().deleteMealPlan(mealPlanId);
+                await GetIt.I<AppDatabase>()
+                    .mealPlansDao
+                    .deleteMealPlan(mealPlanId);
 
                 if (!context.mounted) return;
