+++ lib/screens/recipe.dart
@@ -6,7 +6,7 @@ import 'package:get_it/get_it.dart';
 import 'package:recipes/blocs/recipe/bloc.dart';
 import 'package:recipes/blocs/recipe/events.dart';
 import 'package:recipes/blocs/recipe/state.dart';
-import 'package:recipes/database.dart';
+import 'package:recipes/database/database.dart';
 import 'package:recipes/helpers/add_ingredients_to_groceries.dart';
 import 'package:recipes/helpers/number_formatters.dart';
 import 'package:recipes/models/recipe_detail.dart';
@@ -124,7 +124,9 @@ class SingleRecipeView extends StatelessWidget {
               child: const Text('Yes'),
               onPressed: () async {
                 try {
-                  await GetIt.I<AppDatabase>().deleteRecipe(recipeId);
+                  await GetIt.I<AppDatabase>()
+                      .recipesDao
+                      .deleteRecipe(recipeId);
 
                   if (!context.mounted) return;
